Cleaned up comments and changed severity to Int
- Rewrote comments to be more useful for navigating the codebase
- Changed TrafficReport.severity from Float to Int since SeekBar returns Int

+++ app/src/main/java/dk/itu/moapd/x9/s25134/TrafficReport.kt
@@ -1,8 +1,7 @@
-// TrafficReport.kt
 package dk.itu.moapd.x9.s25134
 
 data class TrafficReport(
     val type: String,
     val description: String,
-    val severity: Float
+    val severity: Int
 )
